Tweaks and upgrade karma/webpack
