+++ karma.conf.cjs
@@ -4,17 +4,26 @@ const webpackConfig = require("./webpack.config.cjs");
 module.exports = function (config) {
   config.set({
     basePath: "",
-    frameworks: ["mocha", "chai"],
-    files: ["test/unit-web/run-web-implementation.test.ts", "test/unit/index-named-exports.test.ts"],
+    frameworks: [
+      "webpack",
+      "mocha",
+      //"chai",
+    ],
+    files: [
+      "test/unit-web/run-web-implementation.test.ts",
+      "test/unit/index-named-exports.test.ts",
+    ],
     exclude: [],
     preprocessors: {
       "test/**/*.ts": ["webpack"],
     },
     webpack: {
       mode: "production",
-      node: webpackConfig.node,
       module: webpackConfig.module,
       resolve: webpackConfig.resolve,
+      experiments: webpackConfig.experiments,
+      optimization: webpackConfig.optimization,
+      stats: {warnings:false},
     },
     reporters: ["spec"],
