+++ package.json
@@ -33,8 +33,8 @@
       ]
     }
   },
-  "module": "./browser",
-  "browser": "./browser",
+  "module": "./lib/index.js",
+  "browser": "./lib/herumi.js",
   "homepage": "https://github.com/chainsafe/bls",
   "author": "ChainSafe Systems",
   "license": "Apache-2.0",
@@ -83,26 +83,27 @@
     "@types/randombytes": "^2.0.0",
     "@typescript-eslint/eslint-plugin": "^4.31.1",
     "@typescript-eslint/parser": "^4.31.1",
+    "buffer": "^6.0.3",
     "chai": "^4.3.6",
     "eslint": "^7.14.0",
     "eslint-plugin-import": "^2.20.1",
     "eslint-plugin-prettier": "^3.1.4",
-    "karma": "^4.4.1",
+    "karma": "^6.3.18",
     "karma-chai": "^0.1.0",
     "karma-chrome-launcher": "^3.1.0",
     "karma-cli": "^2.0.0",
-    "karma-mocha": "^1.3.0",
+    "karma-mocha": "^2.0.1",
     "karma-spec-reporter": "^0.0.32",
-    "karma-webpack": "^4.0.2",
+    "karma-webpack": "^5.0.0",
     "mocha": "^9.2.2",
     "nyc": "^15.0.0",
     "prettier": "^2.1.2",
-    "resolve-typescript-plugin": "^1.1.5",
-    "ts-loader": "^8.3.0",
+    "resolve-typescript-plugin": "^1.2.0",
+    "ts-loader": "^9.2.8",
     "ts-node": "^10.7.0",
     "typescript": "^4.6.3",
-    "webpack": "^4.30.0",
-    "webpack-cli": "^3.3.2"
+    "webpack": "^5.72.0",
+    "webpack-cli": "^4.9.2"
   },
   "resolutions": {
     "mocha": "^9.2.2",
