+++ src/blst-native/index.ts
@@ -1,7 +1,7 @@
 import {SecretKey} from "./secretKey.js";
 import {PublicKey} from "./publicKey.js";
 import {Signature} from "./signature.js";
-import {IBls} from "../interface.js";
+import {IBls} from "../types.js";
 import {functionalInterfaceFactory} from "../functional.js";
 export * from "../constants.js";
