+++ src/herumi/index.ts
@@ -2,7 +2,7 @@ import {SecretKey} from "./secretKey.js";
 import {PublicKey} from "./publicKey.js";
 import {Signature} from "./signature.js";
 import {init, destroy} from "./context.js";
-import {IBls} from "../interface.js";
+import {IBls} from "../types.js";
 import {functionalInterfaceFactory} from "../functional.js";
 
 await init();
