+++ test/unit/secretKey.test.ts
@@ -1,5 +1,5 @@
 import {expect} from "chai";
-import {IBls} from "../../src/interface.js";
+import {IBls} from "../../src/types.js";
 
 export function runSecretKeyTests(bls: IBls): void {
   describe("SecretKey", () => {
